+++ workers/unity/Packages/io.improbable.gdk.buildsystem/WorkerBuilder.cs
@@ -14,14 +14,11 @@ namespace Improbable.Gdk.BuildSystem
 {
     public static class WorkerBuilder
     {
-        private static readonly string PlayerBuildDirectory =
-            Path.GetFullPath(Path.Combine(Directory.GetCurrentDirectory(), EditorPaths.SpatialAssemblyDirectory,
-                "worker"));
+        private static BuildContext? currentContext;
 
-        private static readonly string SpatialAssemblyDirectory =
-            Path.GetFullPath(Path.Combine(Directory.GetCurrentDirectory(), EditorPaths.SpatialAssemblyDirectory));
-
-        private const string BuildWorkerTypes = "buildWorkerTypes";
+        public static BuildContext CurrentContext =>
+            currentContext ?? throw new InvalidOperationException(
+                "CurrentContext is only valid while building a SpatialOS worker.");
 
         /// <summary>
         ///     Build method that is invoked by commandline
@@ -32,73 +29,24 @@ namespace Improbable.Gdk.BuildSystem
             try
             {
                 var args = CommandLineArgs.FromCommandLine();
-                var buildEnvironmentArg = args.GetCommandLineValue("buildEnvironment", "local");
 
-                BuildEnvironment buildEnvironment;
-                switch (buildEnvironmentArg.ToLower())
+                // Parse command line arguments
+                var buildTargetFilter = CommandlineParser.GetBuildTargetFilter(args);
+                var wantedWorkerTypes = CommandlineParser.GetWorkerTypesToBuild(args);
+                var scriptImplementation = CommandlineParser.GetScriptingImplementation(args);
+                var buildEnvironment = CommandlineParser.GetBuildEnvironment(args);
+
+                // Create BuildContext for each worker
+                var buildContexts = BuildContext.GetBuildContexts(wantedWorkerTypes, buildEnvironment, scriptImplementation,
+                    buildTargetFilter);
+
+                if (buildContexts.Count == 0)
                 {
-                    case "cloud":
-                        buildEnvironment = BuildEnvironment.Cloud;
-                        break;
-                    case "local":
-                        buildEnvironment = BuildEnvironment.Local;
-                        break;
-                    default:
-                        throw new BuildFailedException("Unknown build environment value: " + buildEnvironmentArg);
+                    throw new BuildFailedException(
+                        $"Attempted a build with no valid targets!");
                 }
 
-                IEnumerable<BuildTarget> buildTargetFilter = null;
-                var buildTargetFilterArg = string.Empty;
-                if (args.TryGetCommandLineValue("buildTargetFilter", ref buildTargetFilterArg))
-                {
-                    buildTargetFilter = buildTargetFilterArg
-                        .Split(',')
-                        .Select(target =>
-                        {
-                            switch (buildTargetFilterArg.ToLower())
-                            {
-                                case "android":
-                                    return BuildTarget.Android;
-                                case "ios":
-                                    return BuildTarget.iOS;
-                                case "winx86":
-                                    return BuildTarget.StandaloneWindows;
-                                case "win":
-                                    return BuildTarget.StandaloneWindows64;
-                                case "linux":
-                                    return BuildTarget.StandaloneLinux64;
-                                case "macos":
-                                    return BuildTarget.StandaloneOSX;
-                                default:
-                                    throw new BuildFailedException(
-                                        "Unknown build target value: " + buildTargetFilterArg);
-                            }
-                        });
-                }
-
-                var workerTypesArg = args.GetCommandLineValue(BuildWorkerTypes, "UnityClient,UnityGameLogic");
-                var wantedWorkerTypes = workerTypesArg.Split(',');
-
-                ScriptingImplementation scriptingBackend;
-                var wantedScriptingBackend = args.GetCommandLineValue("scriptingBackend", "mono");
-                switch (wantedScriptingBackend)
-                {
-                    case "mono":
-                        scriptingBackend = ScriptingImplementation.Mono2x;
-                        break;
-                    case "il2cpp":
-                        scriptingBackend = ScriptingImplementation.IL2CPP;
-                        break;
-                    default:
-                        throw new BuildFailedException("Unknown scripting backend value: " + wantedScriptingBackend);
-                }
-
-                var buildsSucceeded = BuildWorkers(wantedWorkerTypes, buildEnvironment, buildTargetFilter, scriptingBackend);
-
-                if (!buildsSucceeded)
-                {
-                    throw new BuildFailedException("Not all builds were completed successfully. See the log for more information.");
-                }
+                BuildWorkers(buildContexts);
             }
             catch (Exception e)
             {
@@ -119,10 +67,12 @@ namespace Improbable.Gdk.BuildSystem
             {
                 try
                 {
-                    BuildWorkers(workerTypes, environment);
+                    var buildContexts = BuildContext.GetBuildContexts(workerTypes, environment);
+                    BuildWorkers(buildContexts);
                 }
-                catch (Exception)
+                catch (Exception e)
                 {
+                    Debug.LogException(e);
                     EditorUtility.DisplayDialog("Build Failed",
                         "Build failed. Please see the Unity Console Window for information.",
                         "OK");
@@ -132,11 +82,7 @@ namespace Improbable.Gdk.BuildSystem
             };
         }
 
-        private static bool BuildWorkers(
-            IEnumerable<string> workerTypes,
-            BuildEnvironment buildEnvironment,
-            IEnumerable<BuildTarget> buildTargetFilter = null,
-            ScriptingImplementation? scriptingBackend = null)
+        private static void BuildWorkers(IReadOnlyList<BuildContext> buildContexts)
         {
             var activeBuildTarget = EditorUserBuildSettings.activeBuildTarget;
             var activeBuildTargetGroup = BuildPipeline.GetBuildTargetGroup(activeBuildTarget);
@@ -146,50 +92,32 @@ namespace Improbable.Gdk.BuildSystem
                 const string errorMessage =
                     "Could not find an instance of the SpatialOS Build Configuration.\n\nPlease create one via Assets > Create > SpatialOS > SpatialOS Build Configuration.\n\nIf you already have an instance of the SpatialOS Build Configuration in your project, please open it in the Unity Inspector to force the asset to load and retry the build.";
 
-                if (Application.isEditor)
-                {
-                    EditorApplication.delayCall += () =>
-                    {
-                        EditorUtility.DisplayDialog("Could not find SpatialOS Build Configuration",
-                            errorMessage,
-                            "OK");
-                    };
-                }
+                throw new BuildFailedException(errorMessage);
+            }
 
-                Debug.LogError(errorMessage);
-                return false;
+            if (!Directory.Exists(EditorPaths.PlayerBuildDirectory))
+            {
+                Directory.CreateDirectory(EditorPaths.PlayerBuildDirectory);
             }
 
             try
             {
                 LocalLaunch.BuildConfig();
 
-                var workerResults = new Dictionary<string, bool>();
-                foreach (var wantedWorkerType in workerTypes)
+                foreach (var buildContext in buildContexts)
                 {
-                    var result = BuildWorkerForEnvironment(wantedWorkerType, buildEnvironment, buildTargetFilter, scriptingBackend);
-                    workerResults[wantedWorkerType] = result;
+                    BuildWorkerWithContext(buildContext);
                 }
 
-                var missingWorkerTypes = string.Join(" ", workerResults.Keys.Where(k => !workerResults[k]));
-                var completedWorkerTypes = string.Join(" ", workerResults.Keys.Where(k => workerResults[k]));
-
-                if (missingWorkerTypes.Length > 0)
-                {
-                    Debug.LogWarning(
-                        $"Completed build for {buildEnvironment} target.\n"
-                        + $"Completed builds for: {completedWorkerTypes}\n"
-                        + $"Skipped builds for: {missingWorkerTypes}. See above for more information.");
-                    return false;
-                }
-                else
-                {
-                    Debug.Log($"Completed build for {buildEnvironment} target.");
-                    return true;
-                }
+                Debug.Log($"Completed build for {buildContexts[0].BuildEnvironment} target.");
             }
             catch (Exception e)
             {
+                if (e is BuildFailedException)
+                {
+                    throw;
+                }
+
                 throw new BuildFailedException(e);
             }
             finally
@@ -198,65 +126,12 @@ namespace Improbable.Gdk.BuildSystem
             }
         }
 
-        private static bool BuildWorkerForEnvironment(
-            string workerType,
-            BuildEnvironment buildEnvironment,
-            IEnumerable<BuildTarget> buildTargetFilter,
-            ScriptingImplementation? scriptingBackend = null)
-        {
-            var spatialOSBuildConfiguration = BuildConfig.GetInstance();
-            var environmentConfig = spatialOSBuildConfiguration.GetEnvironmentConfigForWorker(workerType, buildEnvironment);
-
-            var targetConfigs = buildTargetFilter == null
-                ? environmentConfig?.BuildTargets.Where(t => t.Enabled)
-                : environmentConfig?.BuildTargets.Where(t => t.Enabled && buildTargetFilter.Contains(t.Target));
-
-            if (targetConfigs == null || !targetConfigs.Any())
-            {
-                Debug.LogWarning($"Skipping build for {workerType}.");
-                return false;
-            }
-
-            if (!Directory.Exists(PlayerBuildDirectory))
-            {
-                Directory.CreateDirectory(PlayerBuildDirectory);
-            }
-
-            var hasBuildSucceeded = true;
-
-            foreach (var config in targetConfigs)
-            {
-                var buildTargetGroup = BuildPipeline.GetBuildTargetGroup(config.Target);
-                var activeScriptingBackend = PlayerSettings.GetScriptingBackend(buildTargetGroup);
-                try
-                {
-                    if (scriptingBackend != null && config.Target != BuildTarget.iOS)
-                    {
-                        Debug.Log($"Setting scripting backend to {scriptingBackend.Value}");
-                        PlayerSettings.SetScriptingBackend(buildTargetGroup, scriptingBackend.Value);
-                    }
-
-                    hasBuildSucceeded &= BuildWorkerForTarget(workerType, buildEnvironment, config.Target, config.Options);
-                }
-                catch (Exception e)
-                {
-                    throw new BuildFailedException(e);
-                }
-                finally
-                {
-                    PlayerSettings.SetScriptingBackend(buildTargetGroup, activeScriptingBackend);
-                }
-            }
-
-            return hasBuildSucceeded;
-        }
-
         public static void Clean()
         {
             // Delete all but the schema directory where the schema descriptor is placed.
-            if (Directory.Exists(SpatialAssemblyDirectory))
+            if (Directory.Exists(EditorPaths.SpatialAssemblyDirectory))
             {
-                var children = new DirectoryInfo(SpatialAssemblyDirectory).GetDirectories();
+                var children = new DirectoryInfo(EditorPaths.SpatialAssemblyDirectory).GetDirectories();
 
                 foreach (var child in children)
                 {
@@ -273,64 +148,64 @@ namespace Improbable.Gdk.BuildSystem
             }
         }
 
-        private static bool BuildWorkerForTarget(
-            string workerType,
-            BuildEnvironment targetEnvironment,
-            BuildTarget buildTarget,
-            BuildOptions buildOptions)
+        private static void BuildWorkerWithContext(
+            BuildContext buildContext)
         {
-            var spatialOSBuildConfiguration = BuildConfig.GetInstance();
-
-            if (!WorkerBuildData.BuildTargetsThatCanBeBuilt[buildTarget])
-            {
-                var config = spatialOSBuildConfiguration.GetEnvironmentConfigForWorker(workerType, targetEnvironment);
-                var target = config.BuildTargets.First(targetConfig => targetConfig.Target == buildTarget);
-
-                if (target.Required)
-                {
-                    throw new BuildFailedException(
-                        $"Build failed for {workerType}. Cannot build for required {buildTarget} because build support is not installed in the Unity Editor.");
-                }
-
-                Debug.LogWarning($"Skipping {buildTarget} because build support is not installed in the Unity Editor and the build target is not marked as 'Required'.");
-                return false;
-            }
+            var spatialOsBuildConfiguration = BuildConfig.GetInstance();
 
             Debug.Log(
-                $"Building \"{buildTarget}\" for worker platform: \"{workerType}\", environment: \"{targetEnvironment}\"");
+                $"Building \"{buildContext.WorkerType}\" for platform: \"{buildContext.BuildTargetConfig.Target}\", environment: \"{buildContext.BuildEnvironment}\"");
 
-
-            var workerBuildData = new WorkerBuildData(workerType, buildTarget);
-            var scenes = spatialOSBuildConfiguration.GetScenePathsForWorker(workerType);
+            var workerBuildData = new WorkerBuildData(buildContext.WorkerType, buildContext.BuildTargetConfig.Target);
+            var scenes = spatialOsBuildConfiguration.GetScenePathsForWorker(buildContext.WorkerType);
 
             var buildPlayerOptions = new BuildPlayerOptions
             {
-                options = buildOptions,
-                target = buildTarget,
+                options = buildContext.BuildTargetConfig.Options,
+                target = buildContext.BuildTargetConfig.Target,
                 scenes = scenes,
                 locationPathName = workerBuildData.BuildScratchDirectory
             };
 
-            var result = BuildPipeline.BuildPlayer(buildPlayerOptions);
-            if (result.summary.result != BuildResult.Succeeded)
+            var buildTargetGroup = BuildPipeline.GetBuildTargetGroup(buildContext.BuildTargetConfig.Target);
+            var activeScriptingBackend = PlayerSettings.GetScriptingBackend(buildTargetGroup);
+            if (activeScriptingBackend != buildContext.ScriptingImplementation)
             {
-                if (buildTarget == BuildTarget.Android && string.IsNullOrEmpty(EditorPrefs.GetString("AndroidSdkRoot")))
+                PlayerSettings.SetScriptingBackend(buildTargetGroup, buildContext.ScriptingImplementation);
+            }
+
+            try
+            {
+                currentContext = buildContext;
+                var result = BuildPipeline.BuildPlayer(buildPlayerOptions);
+                if (result.summary.result != BuildResult.Succeeded)
                 {
-                    Debug.LogWarning($"Unable to build worker {workerType} for platform Android. " +
-                        $"Ensure you have the Android SDK set inside the Unity Editor Preferences.");
-                    return false;
+                    if (buildContext.BuildTargetConfig.Target == BuildTarget.Android &&
+                        string.IsNullOrEmpty(EditorPrefs.GetString("AndroidSdkRoot")))
+                    {
+                        Debug.LogWarning($"Unable to build worker {buildContext.WorkerType} for platform Android. " +
+                            $"Ensure you have the Android SDK set inside the Unity Editor Preferences.");
+                    }
+
+                    throw new BuildFailedException($"Build failed for {buildContext.WorkerType}");
+                }
+            }
+            catch (Exception e)
+            {
+                if (e is BuildFailedException)
+                {
+                    throw;
                 }
 
-                throw new BuildFailedException($"Build failed for {workerType}");
+                throw new BuildFailedException(e);
             }
-
-            if (buildTarget == BuildTarget.Android || buildTarget == BuildTarget.iOS)
+            finally
             {
-                // Mobile clients can only be run locally, no need to package them
-                return true;
+                currentContext = null;
+                PlayerSettings.SetScriptingBackend(buildTargetGroup, activeScriptingBackend);
             }
 
-            if (buildTarget == BuildTarget.StandaloneOSX)
+            if (buildContext.BuildTargetConfig.Target == BuildTarget.StandaloneOSX)
             {
                 // Unity executable name has changed on MacOS
                 // This is a temp work around our Launcher trying to run the wrong file.
@@ -338,10 +213,12 @@ namespace Improbable.Gdk.BuildSystem
                 CreateLaunchJson(workerBuildData.PackageName, Application.productName);
             }
 
-            var zipPath = Path.Combine(PlayerBuildDirectory, workerBuildData.PackageName);
-            var basePath = Path.Combine(Common.BuildScratchDirectory, workerBuildData.PackageName);
-            Zip(zipPath, basePath, targetEnvironment == BuildEnvironment.Cloud);
-            return true;
+            // Package up standalone builds
+            if (buildContext.BuildTargetConfig.Target != BuildTarget.Android &&
+                buildContext.BuildTargetConfig.Target != BuildTarget.iOS)
+            {
+                Package(workerBuildData.PackageName, buildContext.BuildEnvironment == BuildEnvironment.Cloud);
+            }
         }
 
         private static void CreateLaunchJson(string packageName, string productName)
@@ -363,13 +240,16 @@ namespace Improbable.Gdk.BuildSystem
             }
         }
 
-        private static void Zip(string zipAbsolutePath, string basePath, bool useCompression)
+        private static void Package(string packageName, bool useCompression)
         {
+            var zipPath = Path.Combine(EditorPaths.PlayerBuildDirectory, packageName);
+            var basePath = Path.Combine(Common.BuildScratchDirectory, packageName);
+
             using (new ShowProgressBarScope($"Package {basePath}"))
             {
                 RedirectedProcess.Command(Common.SpatialBinary)
                     .InDirectory(Path.GetFullPath(Path.Combine(Application.dataPath, "..")))
-                    .WithArgs("file", "zip", $"--output=\"{Path.GetFullPath(zipAbsolutePath)}\"",
+                    .WithArgs("file", "zip", $"--output=\"{Path.GetFullPath(zipPath)}\"",
                         $"--basePath=\"{Path.GetFullPath(basePath)}\"", "\"**\"",
                         $"--compression={useCompression}")
                     .Run();